Initial commit - mostly working
Should have been committing on the way but... I didn't think I was
actually going to put this on github. This is a super raw playtesting
application that can be used for any head-to-head comparisons but was
built with TCG playtesting in mind. This is the front end with static
data.

TODO:
* Remove static data
* Create decks
* Connect to API (and create API?)